+++ apps/reviews-api/review/domain/dto/dto.go
@@ -26,3 +26,9 @@ type UserDTO struct {
 	ID       string `json:"id"`
 	Username string `json:"username,omitempty"`
 }
+
+type ReviewDecisionDTO struct {
+	ID             int    `json:"id"`
+	Status         string `json:"status"`
+	RejectedReason string `json:"rejectedReason,omitempty"`
+}
