+++ apps/reviews-api/review/domain/repository/review_repo.go
@@ -5,9 +5,11 @@ import (
 
 	"github.com/stamford-syntax-club/course-compose/reviews/common/utils"
 	"github.com/stamford-syntax-club/course-compose/reviews/review/data/datasource/db"
+	"github.com/stamford-syntax-club/course-compose/reviews/review/domain/dto"
 )
 
 type ReviewRepository interface {
 	GetCourseReviews(context.Context, string, string, *utils.PageInformation) ([]db.ReviewModel, int, error)
 	SubmitReview(context.Context, *db.ReviewModel, string, string) (*db.ReviewModel, error)
+	UpdateReviewStatus(context.Context, *dto.ReviewDecisionDTO) (*db.ReviewModel, error)
 }
